fix: corrected error after moving BaseInput component
